add labels to bipartite
